Restructured file locations
