+++ src/gan_inference.py
@@ -12,15 +12,17 @@ import torch
 from torchvision import transforms
 import math
 import PIL
-from src import saver_and_loader, os_helper
+from utils import saver_and_loader
 from src.configs import ini_parser
 # from src.metrics import score_metrics
-from src.data_load import unnormalize, get_num_classes, create_latent_vector
-from src import create_model
+from data.data_load import unnormalize, get_num_classes, create_latent_vector
+from models import create_model
 
 import os
 import logging
 import argparse
+
+from utils import os_helper
 
 
 def generate_batch_image(ini_config, gan_model, num_images: int):
