+++ src/models/dcgan/dcgan_discriminator.py
@@ -12,7 +12,7 @@ import torch
 import torch.nn as nn
 from torch.nn.utils import spectral_norm
 from src.models.base_discriminator import BaseDiscriminator
-from src import create_model
+from models import create_model
 
 
 class DcganDiscriminator(BaseDiscriminator):
